[aoti] Split custom ops tests
